+++ android/app/src/main/kotlin/com/flaxplayer/flax/MainActivity.kt
@@ -37,6 +37,16 @@ class MainActivity : AudioServiceActivity() {
         }
     }
 
+    private fun requestLocationPermission() {
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(
+                this,
+                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
+                1002
+            )
+        }
+    }
+
     override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
         super.configureFlutterEngine(flutterEngine)
 
@@ -88,6 +98,10 @@ class MainActivity : AudioServiceActivity() {
             when (call.method) {
                 "requestNotificationPermission" -> {
                     requestNotificationPermission()
+                    result.success(true)
+                }
+                "requestLocationPermission" -> {
+                    requestLocationPermission()
                     result.success(true)
                 }
                 "startDownload" -> {